fix(store): cast limit to i64 in get_negative_feedback Entire-Checkpoint: b934bccc5ebd

+++ crates/argus-codelens/src/store.rs
@@ -788,7 +788,7 @@ impl CodeIndex {
             .map_err(|e| ArgusError::Database(format!("failed to prepare feedback query: {e}")))?;
 
         let rows = stmt
-            .query_map(params![limit], |row| row.get(0))
+            .query_map(params![limit as i64], |row| row.get(0))
             .map_err(|e| ArgusError::Database(format!("feedback query failed: {e}")))?;
 
         let mut messages = Vec::new();
